add: alternative solution with MPI_Reduce
